requirements.txt
flask==3.0.2
pyserial==3.5
